fix: parsing protofiles with underscore in fields
This one will exchange protobuf.loadSync with first creating a new
Root() object and using Root.loadSync instead. Apparently the parser of
protobuf converts snake_case to camelCase by default. The
protobuf.loadSync method does not support options to tweak this behavior
so we use the Root method instead.

Since protobufjs goes for an opt-in in keeping the case and converting
to camelCase by default we'll handle this the same and offer a check box
in the configuration node UI.

Fixes #29